Apply the options to the table context menu in toolbar too
